refactor/GormのDBの初期化処理をdbに移行
