+++ internal/infrastructure/db/db_gorm.go
@@ -1,4 +1,4 @@
-package infrastructure
+package db
 
 import (
 	"fmt"
